test(vetting): Wave 8 — split test_3d_shape.h into 3D morphology
10 functions (shared d3shape_GT + test_3shape_feature) split by
registry target_test:
- test_3d_morphology_common.h (git mv from test_3d_shape.h): shared
  map + helper + get_3d_segmented_phantom forward decl.
- test_3d_morphology_regression.h: 8 snapshot shape features.
- test_3d_morphology_matlab.h: 3MESH_VOLUME (matlab/vetted) + the
  MATLAB-derived covariance/eigenvalue math test.

9 registry rows repointed. VERIFIED: full gtest suite 696/696 — no
tests dropped; registry rc=0.
