+++ tests/test_3d_morphology_common.h
@@ -124,116 +124,9 @@ void test_3shape_feature (const std::string& fname, const Nyxus::Feature3D& expe
     ASSERT_TRUE(agrees_gt(atot, d3shape_GT[fname], 10.));
 }
 
-void test_3shape_area() {
-    test_3shape_feature ("3AREA", Feature3D::AREA);
-}
-
-void test_3shape_area2volume() {
-    test_3shape_feature ("3AREA_2_VOLUME", Feature3D::AREA_2_VOLUME);
-}
-
-void test_3shape_compactness1() {
-    test_3shape_feature ("3COMPACTNESS1", Feature3D::COMPACTNESS1);
-}
-
-void test_3shape_compactness2() {
-    test_3shape_feature ("3COMPACTNESS2", Feature3D::COMPACTNESS2);
-}
-
-void test_3shape_meshvolume() {
-    test_3shape_feature ("3MESH_VOLUME", Feature3D::MESH_VOLUME);
-}
-
-void test_3shape_sprericaldisproportion() {
-    test_3shape_feature ("3SPHERICAL_DISPROPORTION", Feature3D::SPHERICAL_DISPROPORTION);
-}
-
-void test_3shape_sphericity() {
-    test_3shape_feature ("3SPHERICITY", Feature3D::SPHERICITY);
-}
-
-void test_3shape_volumeconvhull() {
-    test_3shape_feature ("3VOLUME_CONVEXHULL", Feature3D::VOLUME_CONVEXHULL);
-}
-
-void test_3shape_voxelvolume() {
-    test_3shape_feature ("3VOXEL_VOLUME", Feature3D::VOXEL_VOLUME);
-}
-
-void test_3shape_covmatrix_and_eigenvals() {
-    std::vector<Pixel3> cloud = 
-    {
-        // layout: X, Y, Z, intensity
-        {9,     96,     4,      1000},
-        {26,    55,     89,     1000},
-        {80,    52,     91,	1000 },
-        {3,     23,	80,	1000},
-        {93,    49,	10,	1000},
-        {73,    62,	26,	1000},
-        {49,    68,	34,	1000},
-        {58,    40,	68,	1000},
-        {24,    37,	14,	1000},
-        {46,    99,	72,	1000}
-    };
-
-    double K[3][3];
-    Pixel3::calc_cov_matrix (K, cloud);
-
-    // verdict #1 (covariance matrix)
-    /*
-            producing the ground truth with MATLAB:
-            cloud = [
-                9    96     4,
-                26    55    89,
-                80    52    91,
-                3    23    80,
-                93    49    10,
-                73    62    26,
-                49    68    34,
-                58    40    68,
-                24    37    14,
-                46    99    72] ;
-            K = cov(cloud)
-            >>
-                ans =
-                1.0e+03 *
-                0.9277 - 0.0093 - 0.0601
-                - 0.0093    0.5952 - 0.1913
-                - 0.0601 - 0.1913    1.1933
-    */
-    double gtK[3][3] =
-    {
-        { 0.9277e3,    -0.0093e3,    -0.0601e3 },
-        { -0.0093e3,   0.5952e3,     -0.1913e3 },
-        { -0.0601e3,   -0.1913e3,    1.1933e3 }
-    };
-
-    double tol = 1.0;
-    ASSERT_TRUE (agrees_gt (K[0][0], gtK[0][0], tol));
-    ASSERT_TRUE (agrees_gt (K[0][1], gtK[0][1], tol));
-    ASSERT_TRUE (agrees_gt (K[0][2], gtK[0][2], tol));
-
-    ASSERT_TRUE (agrees_gt (K[1][0], gtK[1][0], tol));
-    ASSERT_TRUE (agrees_gt (K[1][1], gtK[1][1], tol));
-    ASSERT_TRUE (agrees_gt (K[1][2], gtK[1][2], tol));
-
-    ASSERT_TRUE (agrees_gt (K[2][0], gtK[2][0], tol));
-    ASSERT_TRUE (agrees_gt (K[2][1], gtK[2][1], tol));
-    ASSERT_TRUE (agrees_gt (K[2][2], gtK[2][2], tol));
-
-    double L[3];
-    ASSERT_TRUE(Nyxus::calc_eigvals(L, K));
-
-    // verdict #2 (eigenvalues)
-    /*
-    producing the ground truth with MATLAB:
-            L = eig (K)
-            sort(L, 'descend')
-    */
-    double gtL[3] = { 1.2584e3, 0.9202e3, 0.5375e3 };
-    ASSERT_TRUE (agrees_gt(L[0], gtL[0], tol));
-    ASSERT_TRUE (agrees_gt(L[1], gtL[1], tol));
-    ASSERT_TRUE (agrees_gt(L[2], gtL[2], tol));
-}
+// The ten test functions that used to live here have been distributed to their taxonomy homes
+// (registry target_test): the eight regression shape features -> test_3d_morphology_regression.h;
+// 3MESH_VOLUME (matlab/vetted) and the MATLAB-derived covariance/eigenvalue math test ->
+// test_3d_morphology_matlab.h. This header now carries only the shared fixture they include.
