+++ tests/test_all.cc
@@ -39,7 +39,9 @@
 #include "test_imq_regression.h"
 #include "test_3d_nifti.h"
 #include "test_omezarr.h"
-#include "test_3d_shape.h"
+#include "test_3d_morphology_common.h"
+#include "test_3d_morphology_regression.h"
+#include "test_3d_morphology_matlab.h"
 #include "test_3d_gldzm_ibsi.h"
 #include "test_3d_ngldm_ibsi.h"
 #include "test_3d_firstorder_pyradiomics.h"
