+++ tests/vetting/oracle_coverage.csv
@@ -574,15 +574,15 @@ dim,feature,family,status,oracle,agreement,config_recipe,tolerance,current_test,
 3D,3VARIANCE_BIASED,firstorder,regression,,agreed,,,test_3d_feature_coverage.h;test_3d_inten.h,test_3d_firstorder_regression.h,pyradiomics/mirp (native 3D first-order; kurtosis -3),oracle-identified,tracker,test_3d_inten.h ORPHANED (not #included; never run) - left per decision A
 3D,3UNIFORMITY,firstorder,vetted,pyradiomics,agreed,,,test_3d_feature_coverage.h;test_3d_inten.h;test_3d_firstorder_pyradiomics.h,test_3d_firstorder_pyradiomics.h,,,tracker,test_3d_inten.h ORPHANED (not #included; never run) - left per decision A
 3D,3UNIFORMITY_PIU,firstorder,regression,,na,,,test_3d_feature_coverage.h;test_3d_inten.h,test_3d_firstorder_regression.h,pyradiomics/mirp (native 3D first-order; kurtosis -3),oracle-identified,tracker,test_3d_inten.h ORPHANED (not #included; never run) - left per decision A
-3D,3AREA,morphology,regression,,needs_audit,,,test_3d_feature_coverage.h;test_3d_shape.h,test_3d_morphology_regression.h,matlab,promote-after-deepdive,tracker,
-3D,3AREA_2_VOLUME,morphology,regression,,na,,,test_3d_feature_coverage.h;test_3d_shape.h,test_3d_morphology_regression.h,pyradiomics:SurfaceVolumeRatio(3D),oracle-identified,tracker,
-3D,3COMPACTNESS1,morphology,regression,,na,,,test_3d_feature_coverage.h;test_3d_shape.h,test_3d_morphology_regression.h,pyradiomics:Compactness1/Compactness2(3D),oracle-identified,tracker,
-3D,3COMPACTNESS2,morphology,regression,,na,,,test_3d_feature_coverage.h;test_3d_shape.h,test_3d_morphology_regression.h,pyradiomics:Compactness1/Compactness2(3D),oracle-identified,tracker,
-3D,3MESH_VOLUME,morphology,vetted,matlab,agreed,,,test_3d_feature_coverage.h;test_3d_shape.h,test_3d_morphology_matlab.h,,,tracker,
-3D,3SPHERICAL_DISPROPORTION,morphology,regression,,na,,,test_3d_feature_coverage.h;test_3d_shape.h,test_3d_morphology_regression.h,pyradiomics:SphericalDisproportion(3D shape),oracle-identified,tracker,
-3D,3SPHERICITY,morphology,regression,,na,,,test_3d_feature_coverage.h;test_3d_shape.h,test_3d_morphology_regression.h,pyradiomics:Sphericity(3D shape),oracle-identified,tracker,
-3D,3VOLUME_CONVEXHULL,morphology,vetted,matlab,agreed,,,test_3d_feature_coverage.h;test_3d_shape.h,test_3d_morphology_matlab.h,,,tracker,
-3D,3VOXEL_VOLUME,morphology,vetted,matlab,agreed,,,test_3d_feature_coverage.h;test_3d_shape.h,test_3d_morphology_matlab.h,,,tracker,
+3D,3AREA,morphology,regression,,needs_audit,,,test_3d_feature_coverage.h;test_3d_morphology_regression.h,test_3d_morphology_regression.h,matlab,promote-after-deepdive,tracker,
+3D,3AREA_2_VOLUME,morphology,regression,,na,,,test_3d_feature_coverage.h;test_3d_morphology_regression.h,test_3d_morphology_regression.h,pyradiomics:SurfaceVolumeRatio(3D),oracle-identified,tracker,
+3D,3COMPACTNESS1,morphology,regression,,na,,,test_3d_feature_coverage.h;test_3d_morphology_regression.h,test_3d_morphology_regression.h,pyradiomics:Compactness1/Compactness2(3D),oracle-identified,tracker,
+3D,3COMPACTNESS2,morphology,regression,,na,,,test_3d_feature_coverage.h;test_3d_morphology_regression.h,test_3d_morphology_regression.h,pyradiomics:Compactness1/Compactness2(3D),oracle-identified,tracker,
+3D,3MESH_VOLUME,morphology,vetted,matlab,agreed,,,test_3d_feature_coverage.h;test_3d_morphology_matlab.h,test_3d_morphology_matlab.h,,,tracker,
+3D,3SPHERICAL_DISPROPORTION,morphology,regression,,na,,,test_3d_feature_coverage.h;test_3d_morphology_regression.h,test_3d_morphology_regression.h,pyradiomics:SphericalDisproportion(3D shape),oracle-identified,tracker,
+3D,3SPHERICITY,morphology,regression,,na,,,test_3d_feature_coverage.h;test_3d_morphology_regression.h,test_3d_morphology_regression.h,pyradiomics:Sphericity(3D shape),oracle-identified,tracker,
+3D,3VOLUME_CONVEXHULL,morphology,vetted,matlab,agreed,,,test_3d_feature_coverage.h;test_3d_morphology_regression.h,test_3d_morphology_matlab.h,,,tracker,
+3D,3VOXEL_VOLUME,morphology,vetted,matlab,agreed,,,test_3d_feature_coverage.h;test_3d_morphology_regression.h,test_3d_morphology_matlab.h,,,tracker,
 3D,3MAJOR_AXIS_LEN,morphology,regression,,agreed,,,test_3d_feature_coverage.h,test_3d_morphology_regression.h,pyradiomics:MajorAxisLength(3D shape),oracle-identified,tracker,
 3D,3MINOR_AXIS_LEN,morphology,regression,,agreed,,,test_3d_feature_coverage.h,test_3d_morphology_regression.h,pyradiomics:MinorAxisLength(3D shape),oracle-identified,tracker,
 3D,3LEAST_AXIS_LEN,morphology,regression,,agreed,,,test_3d_feature_coverage.h,test_3d_morphology_regression.h,pyradiomics:LeastAxisLength(3D shape),oracle-identified,tracker,
